+++ packages/apps/reputation-oracle/server/src/modules/abuse/abuse-slack-bot.spec.ts
@@ -5,7 +5,7 @@ import { HttpService } from '@nestjs/axios';
 import { createHttpServiceMock } from '../../../test/mock-creators/nest';
 import { SlackConfigService } from '../../config/slack-config.service';
 
-import { AbuseSlackBot } from './abuse.slack-bot';
+import { AbuseSlackBot } from './abuse-slack-bot';
 import { AbuseDecision } from './constants';
 
 const mockHttpService = createHttpServiceMock();
