+++ packages/apps/reputation-oracle/server/src/modules/abuse/abuse.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { HttpModule } from '@nestjs/axios';
-import { AbuseSlackBot } from './abuse.slack-bot';
+import { AbuseSlackBot } from './abuse-slack-bot';
 import { AbuseRepository } from './abuse.repository';
 import { AbuseService } from './abuse.service';
 import { Web3Module } from '../web3/web3.module';
 import { OutgoingWebhookModule } from '../webhook/webhook-outgoing.module';
 import { AbuseController } from './abuse.controller';
-import { AbuseSlackAuthGuard } from './abuse.slack-auth.guard';
+import { AbuseSlackAuthGuard } from './abuse-slack-auth.guard';
 
 @Module({
   imports: [HttpModule, Web3Module, OutgoingWebhookModule],
