+++ packages/apps/reputation-oracle/server/src/modules/nda/nda.controller.ts
@@ -19,7 +19,7 @@ import { RequestWithUser } from '../../common/types';
 import { NDAConfigService } from '../../config/nda-config.service';
 
 import { NDASignatureDto } from './nda.dto';
-import { NDAErrorFilter } from './nda.error.filter';
+import { NDAErrorFilter } from './nda.error-filter';
 import { NDAService } from './nda.service';
 
 @ApiTags('NDA')
